Added endpoint to fetch all cars

+++ src/main/kotlin/com/andrei/finalyearprojectapi/entity/LatLng.kt
@@ -4,8 +4,8 @@ import javax.persistence.*
 
 @Embeddable
 class LatLng(
-    val latitude:Long = 0,
-    val longitude:Long = 0
+    val latitude:Double = 0.0,
+    val longitude:Double = 0.0
 ){
 
 }
